backend/requirements.txt
fastapi
uvicorn
openai>=1.0.0
faiss-cpu
youtube-transcript-api
numpy
python-dotenv
pydantic
